Cria `draw.c` e move código de desenhar para ele
incluindo a função `DesenharControles`
